# Conflicts:
#	game-server/data/handlers/ai/instance/empyreanCrucible/KingConsierdAI.java
#	game-server/data/handlers/ai/instance/empyreanCrucible/MagePreceptorAI.java
#	game-server/data/handlers/ai/instance/empyreanCrucible/PriestPreceptorAI.java
#	game-server/data/handlers/ai/instance/empyreanCrucible/RM1337AI.java
#	game-server/data/handlers/ai/instance/empyreanCrucible/WarriorPreceptorAI.java
#	game-server/data/static_data/npcs/npc_templates.xml

+++ game-server/src/com/aionemu/gameserver/services/craft/CraftService.java
@@ -54,7 +54,7 @@ public class CraftService {
 		int skillLvl = recipetemplate.getSkillpoint();
 		int xpReward = (int) ((0.008 * (skillLvl + 100) * (skillLvl + 100) + 60));
 		xpReward = xpReward + (xpReward * bonus / 100); // bonus
-		int gainedCraftXp = (int) Rates.SKILL_XP_CRAFTING.calcResult(player, xpReward);
+		int gainedCraftXp = Rates.SKILL_XP_CRAFTING.calcResult(player, xpReward);
 		StatEnum boostStat = StatEnum.getModifier(skillId);
 		if (boostStat != null) // there is no boost for morphing (40009)
 			gainedCraftXp *= player.getGameStats().getStat(boostStat, 100).getCurrent() / 100f;
